temp: Remove static from `kernel_device_frames`
This symbol is possibly optimised out by the compiler because it is
static and usually only has a few entries. The seL4CP tool depends on
this symbol existing in the ELF.

The real solution is to have `c_header.py` or another script that
generates the array to also export the kernel device frames in a
format that is easily readable by external tools.