aarch64: makeUserPage: de-dup, use bitfield gen
`makeUserPage` was a bit confusing to understand, and went around the
bitfield-generator's back by modifying the tag bits. As a result, it did
not set bit 58 for 4k pages, meaning a 4k page could not be identified
from an encoded PTE. This version addresses that, and hopefully improves
the repetition.

Signed-off-by: Rafal Kolanski <rafal.kolanski@proofcraft.systems>